Editor: Extract LinkContainer from FormatToolbar
