+++ packages/editor/src/components/rich-text/format-toolbar/index.js
@@ -4,56 +4,17 @@
 import { __ } from '@wordpress/i18n';
 import { Component } from '@wordpress/element';
 import {
-	Fill,
-	IconButton,
-	ToggleControl,
 	Toolbar,
 	withSpokenMessages,
-	Popover,
-	ExternalLink,
 } from '@wordpress/components';
-import { ESCAPE, LEFT, RIGHT, UP, DOWN, BACKSPACE, ENTER, displayShortcut } from '@wordpress/keycodes';
+import { ESCAPE, LEFT, RIGHT, UP, DOWN, BACKSPACE, ENTER } from '@wordpress/keycodes';
 import { prependHTTP } from '@wordpress/url';
 
 /**
  * Internal dependencies
  */
-import PositionedAtSelection from './positioned-at-selection';
-import URLInput from '../../url-input';
-import { filterURLForDisplay } from '../../../utils/url';
-
-const FORMATTING_CONTROLS = [
-	{
-		icon: 'editor-bold',
-		title: __( 'Bold' ),
-		shortcut: displayShortcut.primary( 'b' ),
-		format: 'bold',
-	},
-	{
-		icon: 'editor-italic',
-		title: __( 'Italic' ),
-		shortcut: displayShortcut.primary( 'i' ),
-		format: 'italic',
-	},
-	{
-		icon: 'admin-links',
-		title: __( 'Link' ),
-		shortcut: displayShortcut.primary( 'k' ),
-		format: 'link',
-	},
-	{
-		icon: 'editor-strikethrough',
-		title: __( 'Strikethrough' ),
-		shortcut: displayShortcut.access( 'd' ),
-		format: 'strikethrough',
-	},
-];
-
-// Default controls shown if no `enabledControls` prop provided
-const DEFAULT_CONTROLS = [ 'bold', 'italic', 'strikethrough', 'link' ];
-
-// Stop the key event from propagating up to maybeStartTyping in BlockListBlock
-const stopKeyPropagation = ( event ) => event.stopPropagation();
+import { FORMATTING_CONTROLS } from '../formatting-controls';
+import LinkContainer from './link-container';
 
 /**
  * Returns the Format Toolbar state given a set of props.
@@ -101,7 +62,8 @@ class FormatToolbar extends Component {
 			}
 		}
 		if ( [ LEFT, DOWN, RIGHT, UP, BACKSPACE, ENTER ].indexOf( event.keyCode ) > -1 ) {
-			stopKeyPropagation( event );
+			// Stop the key event from propagating up to maybeStartTyping in BlockListBlock.
+			event.stopPropagation();
 		}
 	}
 
@@ -176,9 +138,11 @@ class FormatToolbar extends Component {
 	}
 
 	render() {
-		const { formats, enabledControls = DEFAULT_CONTROLS, customControls = [], selectedNodeId } = this.props;
+		const { formats, enabledControls = [], customControls = [], selectedNodeId } = this.props;
 		const { linkValue, settingsVisible, opensInNewWindow, isEditingLink } = this.state;
 		const isAddingLink = formats.link && formats.link.isAdding;
+		const isEditing = isAddingLink || isEditingLink;
+		const isPreviewing = ! isEditing && formats.link;
 
 		const toolbarControls = FORMATTING_CONTROLS.concat( customControls )
 			.filter( ( control ) => enabledControls.indexOf( control.format ) !== -1 )
@@ -202,88 +166,26 @@ class FormatToolbar extends Component {
 				};
 			} );
 
-		let linkUIToShow = 'none';
-		if ( isAddingLink || isEditingLink ) {
-			linkUIToShow = 'editing';
-		} else if ( formats.link ) {
-			linkUIToShow = 'previewing';
-		}
-
-		const linkSettings = settingsVisible && (
-			<div className="editor-format-toolbar__link-modal-line editor-format-toolbar__link-settings">
-				<ToggleControl
-					label={ __( 'Open in New Window' ) }
-					checked={ opensInNewWindow }
-					onChange={ this.setLinkTarget } />
-			</div>
-		);
-
 		return (
 			<div className="editor-format-toolbar">
 				<Toolbar controls={ toolbarControls } />
 
-				{ linkUIToShow !== 'none' && (
-					<Fill name="RichText.Siblings">
-						<PositionedAtSelection className="editor-format-toolbar__link-container">
-							<Popover
-								position="bottom center"
-								focusOnMount={ isAddingLink ? 'firstElement' : false }
-								key={ selectedNodeId /* Used to force rerender on change */ }
-							>
-								{ linkUIToShow === 'editing' && (
-									// Disable reason: KeyPress must be suppressed so the block doesn't hide the toolbar
-									/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-									<form
-										className="editor-format-toolbar__link-modal"
-										onKeyPress={ stopKeyPropagation }
-										onKeyDown={ this.onKeyDown }
-										onSubmit={ this.submitLink }>
-										<div className="editor-format-toolbar__link-modal-line">
-											<URLInput value={ linkValue } onChange={ this.onChangeLinkValue } />
-											<IconButton icon="editor-break" label={ __( 'Apply' ) } type="submit" />
-											<IconButton
-												className="editor-format-toolbar__link-settings-toggle"
-												icon="ellipsis"
-												label={ __( 'Link Settings' ) }
-												onClick={ this.toggleLinkSettingsVisibility }
-												aria-expanded={ settingsVisible }
-											/>
-										</div>
-										{ linkSettings }
-									</form>
-									/* eslint-enable jsx-a11y/no-noninteractive-element-interactions */
-								) }
-
-								{ linkUIToShow === 'previewing' && (
-									// Disable reason: KeyPress must be suppressed so the block doesn't hide the toolbar
-									/* eslint-disable jsx-a11y/no-static-element-interactions */
-									<div
-										className="editor-format-toolbar__link-modal"
-										onKeyPress={ stopKeyPropagation }
-									>
-										<div className="editor-format-toolbar__link-modal-line">
-											<ExternalLink
-												className="editor-format-toolbar__link-value"
-												href={ formats.link.value }
-											>
-												{ formats.link.value && filterURLForDisplay( decodeURI( formats.link.value ) ) }
-											</ExternalLink>
-											<IconButton icon="edit" label={ __( 'Edit' ) } onClick={ this.editLink } />
-											<IconButton
-												className="editor-format-toolbar__link-settings-toggle"
-												icon="ellipsis"
-												label={ __( 'Link Settings' ) }
-												onClick={ this.toggleLinkSettingsVisibility }
-												aria-expanded={ settingsVisible }
-											/>
-										</div>
-										{ linkSettings }
-									</div>
-									/* eslint-enable jsx-a11y/no-static-element-interactions */
-								) }
-							</Popover>
-						</PositionedAtSelection>
-					</Fill>
+				{ ( isEditing || isPreviewing ) && (
+					<LinkContainer
+						editLink={ this.editLink }
+						formats={ formats }
+						isEditing={ isEditing }
+						isPreviewing={ isPreviewing }
+						linkValue={ linkValue }
+						onChangeLinkValue={ this.onChangeLinkValue }
+						onKeyDown={ this.onKeyDown }
+						opensInNewWindow={ opensInNewWindow }
+						selectedNodeId={ selectedNodeId }
+						setLinkTarget={ this.setLinkTarget }
+						settingsVisible={ settingsVisible }
+						submitLink={ this.submitLink }
+						toggleLinkSettingsVisibility={ this.toggleLinkSettingsVisibility }
+					/>
 				) }
 			</div>
 		);
